feat: 강의 진행 상태에서 수강 신청 가능하도록 수정
