+++ README.md
@@ -54,8 +54,8 @@ validate도 마찬가지
 ## Step4
 
 ### TODO
-- [ ] 강의가 진행 중인 상태에서도 수강신청이 가능해야 한다.
-- [ ] 강의 진행 상태(준비중, 진행중, 종료)와 모집 상태(비모집중, 모집중)로 상태 값을 분리해야 한다.
+- [x] 강의가 진행 중인 상태에서도 수강신청이 가능해야 한다.
+- [x] 강의 진행 상태(준비중, 진행중, 종료)와 모집 상태(비모집중, 모집중)로 상태 값을 분리해야 한다.
 - [ ] 강의는 하나 이상의 커버 이미지를 가질 수 있다.
 - [ ] 강사는 수강신청한 사람 중 선발된 인원에 대해서만 수강 승인이 가능해야 한다.
 - [ ] 강사는 수강신청한 사람 중 선발되지 않은 사람은 수강을 취소할 수 있어야 한다.
