+++ src/main/java/nextstep/courses/infrastructure/JdbcSessionRepository.java
@@ -5,7 +5,7 @@ import nextstep.courses.domain.Image;
 import nextstep.courses.domain.Image.Dimension;
 import nextstep.courses.domain.ImageType;
 import nextstep.courses.domain.Session;
-import nextstep.courses.domain.Session.SessionDuration;
+import nextstep.courses.domain.SessionDuration;
 import nextstep.courses.domain.SessionPayType;
 import nextstep.courses.domain.SessionRepository;
 import nextstep.courses.domain.SessionState;
